requirements.txt
psutil~=5.8.0
ffpyplayer~=4.3.2
dill~=0.3.4
paramiko~=2.8.0
numpy~=1.20.3
pygame~=2.1.0
opencv-python~=4.5.3.56
geojson~=2.5.0
requests~=2.26.0
Pillow~=8.3.1
Pint~=0.18
atmos~=0.2.6
urllib3~=1.26.6
django-os-geocoder~=0.1.6
win10toast~=0.9
pyserial~=3.5
geocoder~=1.38.1